+++ tests/content_type_field_types_test.py
@@ -14,7 +14,8 @@ from contentful.content_type_field_types import (
     LinkField,
     ArrayField,
     ObjectField,
-    RichTextField
+    RichTextField,
+    ResourceLinkField
 )
 from contentful.resource import Link
 
@@ -241,3 +242,10 @@ class RichTextFieldTest(TestCase):
             "nodeType": "document",
             "content": []
         })
+
+
+class ResourceLinkFieldTest(TestCase):
+    def test_resource_link_field(self):
+        resource_link_field = ResourceLinkField()
+
+        self.assertEqual(resource_link_field.coerce('foo'), 'foo')